pyproject.toml
[project]
name = "resume-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-google-vertexai>=2.0.27",
    "python-docx>=1.2.0",
    "uvicorn>=0.35.0",
]
